Move sync code to main
It's almost all pico/picosystem-specific or debug code
