+++ CMakeLists.txt
@@ -57,7 +57,6 @@ picosystem_executable(
     engine/render_lighting.cpp
     engine/render_model.cpp
     engine/render_rasterize.cpp
-    engine/render_sync.cpp
     engine/render_triangle.cpp
 
     game/logic_day_night_cycle.cpp
